Crate bitmap_font

This version is for embedded-graphics 0.7. The latest release works with embedded-graphics 0.6.

This crate provides bitmap fonts for the embedded-graphics crate. Those don’t only look better than the built-in fonts by using the good-looking Tamzen font over a font that renders . like a +, but also allow scaling fonts by pixel-doubling them, giving you two font sizes for the flash size requirements of the smaller one.

Structs

BitmapFont
Stores the font bitmap and some additional info for each font.

Constants

FONT_5x9
Tamzen 5x9 Font.

FONT_5x9_BOLD
Tamzen 5x9 Bold Font.

FONT_6x12
Tamzen 6x12 Font.

FONT_6x12_BOLD
Tamzen 6x12 Bold Font.

FONT_7x13
Tamzen 7x13 Font.

FONT_7x13_BOLD
Tamzen 7x13 Bold Font.

FONT_7x14
Tamzen 7x14 Font.

FONT_7x14_BOLD
Tamzen 7x14 Bold Font.

FONT_8x15
Tamzen 8x15 Font.

FONT_8x15_BOLD
Tamzen 8x15 Bold Font.

FONT_8x16
Tamzen 8x16 Font.

FONT_8x16_BOLD
Tamzen 8x16 Bold Font.

FONT_10x18
Pixel-double version of FONT_5x9.

FONT_10x18_BOLD
Pixel-double version of FONT_5x9_BOLD.

FONT_10x20
Tamzen 10x20 Font.

FONT_10x20_BOLD
Tamzen 10x20 Bold Font.

FONT_12x24
Pixel-double version of FONT_6x12.

FONT_12x24_BOLD
Pixel-double version of FONT_6x12_BOLD.

FONT_14x26
Pixel-double version of FONT_7x13.

FONT_14x26_BOLD
Pixel-double version of FONT_7x13_BOLD.

FONT_14x28
Pixel-double version of FONT_7x14.

FONT_14x28_BOLD
Pixel-double version of FONT_7x14_BOLD.

FONT_16x30
Pixel-double version of FONT_8x15.

FONT_16x30_BOLD
Pixel-double version of FONT_8x15_BOLD.

FONT_16x32
Pixel-double version of FONT_8x16.

FONT_16x32_BOLD
Pixel-double version of FONT_8x16_BOLD.

FONT_20x40
Pixel-double version of FONT_10x20.

FONT_20x40_BOLD
Pixel-double version of FONT_10x20_BOLD.